api/requirements.txt
python-dotenv==0.15.0
Flask==1.1.2
wolframalpha==4.1.0
flask-cors==3.0.9